split rotation tests into indep files to speed up unit test time
